+++ tests/Frontend/e2e/RoomsViewGeneral.cy.js
@@ -543,10 +543,9 @@ describe("Room View general", function () {
     );
     cy.get("#invitationCode").should("have.value", "508-307-005");
 
+    // Copy invitation message
     cy.get('[data-test="room-copy-invitation-button"]').click();
-
-    cy.checkToastMessage("rooms.invitation.copied");
-
+    cy.checkToastMessage("rooms.invitation.copied_message");
     cy.window().then((win) => {
       win.navigator.clipboard.readText().then((text) => {
         expect(text).to.eq(
@@ -554,6 +553,26 @@ describe("Room View general", function () {
             Cypress.config("baseUrl") +
             "/rooms/abc-def-123\nrooms.invitation.code: 508-307-005",
         );
+      });
+    });
+
+    // Copy room link
+    cy.get('[data-test="room-share-button"]').click();
+    cy.get('[data-test="room-invitation-copy-link-button"]').click();
+    cy.checkToastMessage("rooms.invitation.copied_url");
+    cy.window().then((win) => {
+      win.navigator.clipboard.readText().then((text) => {
+        expect(text).to.eq(Cypress.config("baseUrl") + "/rooms/abc-def-123");
+      });
+    });
+
+    // Copy room access code
+    cy.get('[data-test="room-share-button"]').click();
+    cy.get('[data-test="room-invitation-copy-code-button"]').click();
+    cy.checkToastMessage("rooms.invitation.copied_code");
+    cy.window().then((win) => {
+      win.navigator.clipboard.readText().then((text) => {
+        expect(text).to.eq("508-307-005");
       });
     });
   });
@@ -1792,6 +1811,63 @@ describe("Room View general", function () {
     cy.url()
       .should("include", "/404")
       .should("not.include", "/rooms/abc-def-123");
+  });
+
+  it("auto-reload", function () {
+    cy.intercept("GET", "api/v1/currentUser", {});
+    cy.fixture("config.json").then((config) => {
+      config.data.room.refresh_rate = 60;
+      cy.intercept("GET", "api/v1/config", {
+        statusCode: 200,
+        body: config,
+      });
+    });
+    cy.interceptRoomFilesRequest();
+    cy.fixture("room.json").then((room) => {
+      room.data.allow_membership = true;
+      room.data.current_user = null;
+
+      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
+        statusCode: 200,
+        body: room,
+      }).as("roomRequest");
+    });
+    cy.clock();
+    cy.visit("/rooms/abc-def-123");
+    cy.wait("@roomRequest");
+    cy.contains("Meeting One").should("be.visible");
+
+    // Wait more than 60 seconds (due to reload randomness) for a auto-reload
+    cy.tick(100000);
+    cy.get("@roomRequest.all").should("have.length", 2);
+  });
+
+  it("auto-reload disabled on error", function () {
+    cy.intercept("GET", "api/v1/currentUser", {});
+    cy.fixture("config.json").then((config) => {
+      config.data.room.refresh_rate = 60;
+      cy.intercept("GET", "api/v1/config", {
+        statusCode: 200,
+        body: config,
+      });
+    });
+    cy.intercept("GET", "api/v1/rooms/abc-def-123", {
+      statusCode: 403,
+      body: {
+        message: "guests_not_allowed",
+      },
+    }).as("roomRequest");
+
+    cy.clock();
+    cy.visit("/rooms/abc-def-123");
+    cy.wait("@roomRequest");
+
+    // Check that the error message is shown
+    cy.contains("rooms.only_used_by_authenticated_users").should("be.visible");
+
+    // Wait more than 60 seconds (due to reload randomness) for no auto-reload
+    cy.tick(100000);
+    cy.get("@roomRequest.all").should("have.length", 1);
   });
 
   it("reload with errors", function () {